refactor: misc refactoring
